+++ src/number/clamp/define.js
@@ -1,0 +1,3 @@
+import { defineInstance } from '../../core/number';
+import clamp from './';
+defineInstance('clamp', clamp);