+++ src/number/index.js
@@ -1,6 +1,7 @@
 export { default as abbr } from './abbr';
 export { default as abs } from './abs';
 export { default as ceil } from './ceil';
+export { default as clamp } from './clamp';
 export { default as exp } from './exp';
 export { default as floor } from './floor';
 export { default as format } from './format';