+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v3.0.0
 - Removed `base` param of `Number#pad` in favor of `Number#toHex`.
 - Removed trigonometric aliases sin, cos, etc in favor of native Math.
 - Changed `Number#metric` placeholder token from `_` to `-`.
+- Removed `Number#cap` in favor of `Number#clamp` with one argument.
 
 TODO: